Fix wrong live stream duration
StageFright has a default duration of 1 hour for all streams that
do not report a duration. Most streams that do not have a duration
are live streams, so the duration is not relevant. This fix
changes the default value to 0 seconds instead, so that the user
does not think that a live stream will end in 1 hour.

Change-Id: If80fb9d6eac7e1ccc751c0979d09526f65189217