Update patch set 1
Patch Set 1:

Yes - this will be an endless game. for instance, if there are multiple, which 'stts' should we use? the first one, the second one, the third one, ...?

I am not saying that we are not aware of the standard non-conforming issue, but this is the first time that we have the request for supporting multiple stts boxes, and this must be fixed by the content creator (or creating software).

Patch-set: 1