Update patch set 1
Patch Set 1:

Yes, I tested it before, but not thoroughly enough. It works fine even in valgrind on x86, but on arm, if the width isn't a multiple of 8, one ends up with segfaults (probably due to trying to read from unaligned addresses). If the height isn't a multiple of 16, it works pretty well.

However, additionally, I inspected the visual quality on x86 closer, and it seemed that the edge DCTs would overread from following data (which didn't trigger warnings in valgrind for some reason), leading to some visual artefacts. This could of course be handled as well, but it did turn out to be slightly more complicated than I initially thought, so I'll abandon this case for now.

Patch-set: 1