Update patch set 1
Patch Set 1:

As I said before, the OMX wrapper already prepends the startcodes for the codecconfig data, so if startcodes are added in the PV encoder, the handling of the codecconfig data needs changing as well.

Currently, the PV encoder API in itself is a lower level API, agnostic of startcodes, while the startcodes are required on the OMX level and thus are added in the OMX wrapper. If you change this, you basically redesign the whole PV encoder API.

Patch-set: 1