Update patch set 1
Patch Set 1:

I also think the log is meaningless, Can we view this log?

LOG_ALL_BUT_BYTES is introduced in commit:
http://git.chromium.org/gitweb/?p=chromium/src/net.git;a=commitdiff;h=a132d93b25b680d6ddc1ec9ff3c9988c9e17770e

before this commit, LOG_ALL will not log transferred bytes.so I'd rather believe that someone did not aware this change or the real meaning of LOG_ALL when he implemented this file.

Patch-set: 1