Update patch set 1
Patch Set 1:

This patch was originally
https://android-review.googlesource.com/#/c/31970/1

Please see the review comments there.

The crash was observed in a CTS run on honeycomb and not noticed since.  Whether or not the CTS test was correctly applied for that system, the result should not be a null string deref crash in stagefright.  How the run got as far as the failed call to isCameraColorFormatSupported() is unexplained.  Perhaps current stagefright now catches such problems earlier.  Or perhaps not.

Bhanu, who diagnosed this problem, is no longer at Mips and so cannot help with confirming whether this problem still exists.

Patch-set: 1