Read track language information from Matroska files
Add track language as kKeyMediaLanguage to support user selection
of audio and subtitle tracks based on language.

According to the Matroska standard language codes can be either
the 3 letters bibliographic ISO-639-2 form (like "fre" for french),
or a language code mixed with a country code for specialities in
languages (like "fre-ca" for canadian french).
Android can only return a language code in ISO-639-1 or ISO-639-2
so any country code is truncated away by this patch.

Test: check language info in multiple audio track matroska file

Change-Id: I6b95bbcb4d3f99b87ea01a22849ba2560a85829a