Update patch set 1
Patch Set 1: Do not submit

I do not see any reason that the stopping logic needs to be changed. If camera takes too long to send frames, it is a problem in the HAL. In addition, this patch may introduce hidden or corner cases that are difficult to track down.

Patch-set: 1
Label: Code-Review=-2