Update patch set 1
Patch Set 1:

There's a theme here that I've seen in quite a few commits. We have been trying to play questionable streams (our customers demand it) and have been fixing Google code to "do the right thing" with garbage-in data. This involves either looking a bit harder for the right data or skipping to the next frame, etc.

In each case, the commit has been rejected with "we don't want to process bad data". I get it... bad data is inconvenient; it is however a fact of life. I believe that Android will have more satisfied customers by fixing these types of problems than other smart phones that don't fix them.

To be sure, we can keep them as local fixes in our tree... and will likely do so... but IMO it is better to float the boat higher for everyone.

... ok... end of my editorial.

Patch-set: 1