Update patch set 2
Patch Set 2:

Hi Bruce,
Commit 46127 handles more different codecs
and provides a framework to add support for some codecs
that may not be present on all platforms.

Even if all codecs are not relevant for ASOP I think it
is good to provide some support for vendors that have them.

It is also plausible that there are codecs out there that
handles several types of content, but they should be
clearly separated in the parser.

Patch-set: 2