Update patch set 4
Patch Set 4:

Hi,
I was wondering if you have a follow-up patch? Thanks!

The patch is close to what I have in mind for DrmBuffer, but missing a couple of points. Mostly, the copy constructor and the copy assignment operator is still not good enough and missing the self check. That could lead to deallocing memory by accident.

Did not review other parts of the patch yet.

Patch-set: 4