Update patch set 3
Patch Set 3:

Yes it's intended. If I remember correctly it was becase we were not allowed to create the file in the test cases. But we are allowed to modify the file so when converting the .dm file we will write the result in the empty .fl file.

Patch-set: 3
Reviewer: Gerrit User 1004910 <1004910@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0