Update patch set 2
Patch Set 2:

I modified the patch according to your comments.
However, it seems that there is another problem.

BpMediaMetadataRetriever is a binder proxy class. So, If extractMetadata() function keeps copies of all the metadata in a keyed vector, the size of the keyed vector can increase infinitely because BpMediaMetadataRetriever doesn't know when it can clear the keyed vector.

I think the best way to solve this issue is changing interface of extractMetadata() function to return String8 instead of char*.

But I have no idea whether changing interface is possible at this time.

Patch-set: 2