Update patch set 1
Patch Set 1:

Your fix is good, but it will cause a merge conflict because we have an alternative fix for this, which was already submitted to our internal branch (but is not yet on AOSP).  Can you please upload this version instead:

         } else {
             toWrite = audioBuffer.size;
             memcpy(audioBuffer.i8, src, toWrite);
-            src += toWrite;
         }
+        src += toWrite;
         userSize -= toWrite;
         written += toWrite;


Then your version should be able to be submitted and will result in a trivial merge.

Patch-set: 1
Label: Code-Review=0