Update patch set 3
Patch Set 3:

(3 comments)

General comment on the word "Beats" in filename and source code:  typical attribution for Android AOSP is via the git commit log, not by the filenames, comments, or symbol names in the source code. There are exceptions especially in external/, but we find that attribution via git log only helps keep the code cleaner and easier to read.   In this particular case, I think that the word "Beats" is minor, and it will probably be fine to leave it as is for now (but I'll check with our AOSP lead to confirm).   But for future contributions, please keep this in mind.

Patch-set: 3