Update patch set 3
Patch Set 3:

The existing resamplers (linear, cubic, high quality,and very high quality) are written in such a way that the quality is implicit.

With a polyphase resampler, using multiple qualities in effect allows variability of quality with complexity (linearly).

Adding the API first without the polyphase resampler doesn't make sense.

Adding our resampler first without the API could make sense if we restrict to a single quality/complexity.

If that is the case, then we just choose one.  Is this what you are asking?  If so, then we will need advice on which one to choose.

Patch-set: 3