Update patch set 3
Patch Set 3:

FWIW, I tried modifying EncodeDecodeTest from CTS to run on a normal device as a normal app, and on Nexus7 on 4.3, it passes the tests using buffers as input (when doing the equivalent to this patch within EncodeDecodeTest). The tests that use a surface for input do fail with "E/OMXNodeInstance: createInputSurface requires AndroidOpaque color format" - fixing that does feel out of my league at the moment though.

So in short, this patch should make this encoder at least able to handle half of EncodeDecodeTest - but please do test for yourself, my testing of this right now is only on a unmodified retail device.

Patch-set: 3