Update patch set 4
Patch Set 4:

(1 comment)

> 1. make doesn't work with Neon

This is probably due to lack of #if at line 35 of .cpp

> 2. code cleanup 

Not relevant until we can address the performance.
We need to show significantly better performance at
a given quality level, or significantly better quality
at a given performance level, or else this is moot.
So let's focus on performance now and note code cleanup issues for later.

> 3. code optimization (intrinsic vs asm)

This should be the focus now.

> If the intrinsic library is not making properly, wouldn't this be expected (at least for now)?

I did a build for a device which has NEON, and I see NEON in the disassembled output.  It just doesn't dominate and thus I suspect the code is not taking best use of NEON.

AOSP build is for a device without NEON, hence the build error, see #1.

Patch-set: 4