Update patch set 3
Patch Set 3:

Ping Lajos, any progress on testing and approving this change? It should be pretty straightfoward to test. Without this change, no EncodeDecodeTests pass with this encoder, with this change the buffer input tests should pass (on most decoders at least, except for the SW decoder). With this and https://android-review.googlesource.com/63201, buffer input tests should pass with the SW decoder as well.

Patch-set: 3