Update patch set 2
Patch Set 2:

Lajos, Glenn, ping. This patch can easily be tested with EncodeDecodeTest by using the SW H264 decoder, trying to decode the output from the SW H264 encoder. Previously it fails on many frames due to the SW decoder being very strict with adhering to the standard. This makes the encoder adhere as strictly to the standard.

Patch-set: 2
Label: Verified=0