Skip ID3v2 header when sniffing FLAC format media file
If a FLAC file contains an ID3v2 header the current FLAC extractor will
not recognize it as a FLAC file. The FLAC reference decoder will locate
and one skip ID3v2 header.

To be able to sniff more FLAC files skip ID3v2 header so that the FLAC
extractor can locate the signature word to recognize such variants of
FLAC files.

Bug: 69404904
Test: manual
Change-Id: I0276f1cf8a31c734207684c5507bc510bec0a00c