Update patch set 1
Patch Set 1: -Code-Review

On 2nd look, the change description seems incorrect.
The only variable in the log is mediaTimeUs, however that is still a valid name and wouldn't lead to compiler failure.
Functionally, the code seems identical to what it was before, except it now recalculates the value to print instead of using the value that was calculated a few lines up.

Patch-set: 1