Update patch set 1
Patch Set 1:

yeah,we are developing a wifi display apk based on your deleted reference code(DirectRender). Then we found latency is too large. So we add a dropping frame algorithm which maybe cause this. Of course, we can add a check step to avoid this issue, but considering here you have added a condition check. so why not make it more strict and then app can just ignore check and app codes will be more clean and efficiency.

Patch-set: 1