Update patch set 1
Patch Set 1:

Sorry for my abysmal response time on this CL :-(

I've finally gotten around to investigating this, and eventually realized that it would never make sense to resample on the fast path ... originally my reason for making it a compile-time option (and disabled by default) was that I was worried about the CPU time for resampling, and that FastMixer thread should run very often but with very low CPU usage per run.   Now, looking back, I realize another reason to avoid resampling is that it is difficult to wake up the client thread (which is also SCHED_FIFO) at regular periodic intervals when resampling occurs, especially for "weird" ratios like 44.1/48 or 48/44.1.  (24/48 or 44.1/22.05 are easier).    So my leaning now is to remove the code that allows resampling on fast path.   If you really want to include support in your own builds, how about if you add an #ifdef around this new section of code you just added?

p.s. I'm also going to copy Raph to get his opinion on this.

Patch-set: 1