Update patch set 1
Patch Set 1:

Ah, thanks for the clarification.  I also want to discuss this with Raph when he's available ... he has had a lot of insights about the impact of sample rate conversion and buffer size configuration in the past, and may be able to help us again here when he has time.

Patch-set: 1