Update patch set 1
Patch Set 1:

You can use LOCAL_MODULE_RELATIVE_PATH to set a path that is relative to the default path (for example soundfx/).  If you need an explicit path you would have to create two build rules, one for the first architecture and one for the second.  Or we would need to add more features to the build system.  So far I haven't found anywhere that needs that.

Patch-set: 1