Update patch set 2
Patch Set 2:

Do you have a test case (sample data that will result in the i==7 and a crash)?  Generally we look for a fail -> pass transition on a test case.  If you could please add a URL to a sample test stream that will result in the 7, it would be great.  Please make sure there are no copyright / license restrictions.  Thanks

Patch-set: 2