Update patch set 4
Patch Set 4:

Can you please split this into one CL for AudioPolicyService, which should be able to be +2 quickly, and a second empty CL (or comment to the first CL) which is a shell/sed script that makes the necessary changes?  There are so many lines changed here that it will be easier for me to review the shell/sed script than to review the affected files.  Thanks

Patch-set: 4
Reviewer: Gerrit User 1004395 <1004395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0