Update patch set 1
Patch Set 1:

I'm not sure I'm convinced about this. ZSL should be a copy of result->request as much as possible, I am not too fond of this magically mutate some of the result fields part.

What is this trying to accomplish that can't be done by setting extra vendor-specific keys?

Patch-set: 1