Update patch set 1
Patch Set 1:

This patch just allows to selectively revert the earlier changes - so that you can see the differences.  The logic, while looking similar is distinct between the PTS and the non-PTS use case.  In non-PTS we try to schedule to be right on time, in non-PTS, we just want to be somewhere within 20-30ms ahead of the time to display (and we allow greater granularity when playing content)

Patch-set: 1
Label: Verified=0