Update patch set 1
Patch Set 1:

> (2 comments)
 > 
 > is there a reason why the existing encryption mechanism is not
 > suitable?
May be the description is not clear, it should be that if the encoder doesn’t support HDCP encryption, then use existing support. Do you see any issue with this?
 > (2 comments)
 > 
 > is there a reason why the existing encryption mechanism is not
 > suitable?

we’d prefer to not use encryptNative() or encrypt() for our platform. The encoder output has it encrypted already, it will be very hard to use encryptNative() or encrypt().

Patch-set: 1