Update patch set 1
Patch Set 1: Code-Review+1

This looks way more robust to me, than the other patch.

Yes, naming something ULong when it specifically should be 32 bit seems wrong, but perhaps someone should have thought of that 16 years ago when this code was written. This is the patch that most likely will work as intended for all the cases that aren't tested right now. If you are irked by the fact that ULong really isn't long, then the type can be renamed later as a separate step.

Patch-set: 1
Label: Code-Review=+1