Update patch set 1
Patch Set 1: Code-Review-1

I don't think this patch can be submitted in its current state.  The reason is that ULong is actually defined in two separate places in the code, once in mp4def.h and once in mp4enc_api.h.  The actual type of ULong in a given file depends on which of the above header files are included, and in the case they are both included, on which file is included first.

Needless, to say having two separate definitions for ULong that do not agree is probably not really what we want.  My guess is that the introduction of this discrepancy could mean that callers and callees are using different versions of ULong

Patch-set: 1
Label: Code-Review=-1