Update patch set 2
Patch Set 2:

As Martin Storsjö indicated earlier, if Google engineers agree with changing Ulong to uint32_t, please feel free to change all ULong to uint32_t in another patch. 

However, personally, I would leave ULong as is, for following reasons 
(a) Original authors assumed ULong to be 32-bit, for reasons we can only speculate.  
(b) There are 406 occurrences of ULong in the code. Prior to changing it, we will need to examine each usage and determine if the usage was unsigned long or just uint32_t.

Patch-set: 2