Update patch set 2
Patch Set 2:

Also, to add in my experience of similar encoder code bases - most of them use ULong just because they (erroneously) chose such a convention, and used it throughout the code for anything that should be 32 bit.

The hard part is that whichever way this is fixed, one should ideally run a full test of the codec - not just checking that it doesn't crash, but making sure the output is correct, in all corner cases as well, not only for the simple cases. And when such a testsuite isn't available, one can only try to pick the solution with the least risk.

And I would say that switching all of the ULong to 32 bit, except for the few cases where it's used for pointer arithmetics (which are easily identifyable), is the option with least risk of breaking anything in a less-tested codepath.

Patch-set: 2
Reviewer: Gerrit User 1002068 <1002068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0