Update patch set 1
Patch Set 1:

Yes, I tested this code and the generated video file is valid. Note that the actual ESDS descriptor is never used. If you search for the usages of kKeyESDS you'll see that it's always used like this:

ESDS esds(data, size);
esds.getCodecSpecificInfo(&data, &size);

and then only CSD data is used.

The cleanest change would be to replace kKeyESDS and kKeyAVCC keys with kKeyCSD which would contain only CSD without wrapping it in ESDS/AVCC which are later ignored anyway. This would require some more work so I think that my change is a good interim solution.

Patch-set: 1