Update patch set 1
Patch Set 1:

I agree it is logically better to close it in the same function :-).  Thanks for the explanation on the repro case, that makes sense now.  Let me put this under review for a possible future release (might be a month before I can get back to this).  Thanks!

Patch-set: 1