Update patch set 1
Patch Set 1:

Thank you for your explanation, now it makes sense.

Andy and I also discussed it further, and we agree with your fix and also want to apply it to another case for binderDied.

Next I need to confirm that I can get the failure by testing, and then that the problem goes away after the fix. This will take me a day or two.

Patch-set: 1