Using requested video scaling mode in NuPlayer
In comparison to AwesomePlayer, NuPlayer has lack of setting
the video scaling mode flag on event kWhatVideoSizeChanged.
That causes the issues with wrong video screen ratio, because
ACodec has hardcoded flag which was the last one set up for
particular native window object, even that flow goes back to NuPlayer.

Change-Id: I1fe28e64d58cdec0cda596eb2f6640ab9be7fd21