Update patch set 2
Patch Set 2:

> > Thanks for cleaning this up!
 > > Out of curiosity, can you tell me what you're using this
 > interface
 > > for?
 > 
 > Not actually working on Android, to be honest. :)
 > I'm a Firefox OS engineer and trying to implement getUserMedia()
 > with camera2 API.

Yes, I figured given your address, and since you control the whole OS stack for Firefox OS, I'll only mildly warn you that (as you can clearly tell) the native client-side interface here is wholly untested.  The public Java API talks directly through Binder with the camera service, so only the service-side native binder interface is used today.

This binder layer is also likely to change in the future, as there's some overdue cleanup due for the camera2 API guts.  But that may not happen all that soon. But this is considered an entirely internal interface, so no guarantees about it staying as it is.

Patch-set: 2