Update patch set 4
Patch Set 4:

For this change, we probably want to use our own fixes to fix similar issues.

I'm working on a patch to move PlaylistFetchers to a separate thread. We'll get rid of the 10sec polling , but keep the 1 sec polling which (is for DownSwitching only currently). Then we'll do centralized up/down switching control there.

We want to put the buffering logic in LiveSession and do a regular polling, so that we can apply similar logic to HLS as that for other type of source (eg. GenericSource).

Patch-set: 4
Reviewer: Gerrit User 1039918 <1039918@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0