Update patch set 2
Patch Set 2:

mSlots[item.mBuf].mGraphicBuffer is only updated to item.mGraphicBuffer if item.mGraphicBuffer is not NULL, ie when it's acquired for the first time. After that the mGraphicBuffer pointer is supposed to be cached in mSlots.

If it somehow passed a NULL mGraphicBuffer to passMetadataBuffer, that means the slot is used before it ever acquired any buffer. This seems a bug somewhere else.

After the first acquire, item.mGraphicBuffer from acquireBuffer will always be NULL for a slot id. all reads will fail after the first one.

Patch-set: 2