Update patch set 4
Patch Set 4:

The idea we had to resolve this problem is to have getOutput() not take a reference on the output stream (that is close it when done) whereas getOutputForAttr() will keep the output open and increment ref count.
The problem as with your change is that it creates a lot of churn because direct outputs are open/closed sevreal times each time you create an AudioTrack on a direct output.
I still don't have the ideal solution for that.

Patch-set: 4