Update patch set 4
Patch Set 4:

what I am saying is that if you call releaseOutput() after getOutput() like you are proposing then you will effectively open and close an output stream each time.
This does not happen if you don't call releaseOutput() as the output stream remains opened.

Your last proposal brings us back to the output cache we had a few releases back in AudioSystem which we removed because it was complex and not reliable.

I prefer to change the contract for getOutput() in AudioPolicyManager implementation and not increment the ref count on the output stream in getOutput() but only in getOutputForAttr().

Also we need to make an inventory of all callers to getMinFrameCount(), getOutputSamplingRate() etc.. and try to suppress them. It is most of the times irrelevant to call those before track construction as the result is very often meaningless as the final parameters applicable to a particular use case are only determined once the track is constructed.

Patch-set: 4
Reviewer: Gerrit User 1003259 <1003259@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0