Update patch set 5
Patch Set 5:

Hi Yin-Chia,

The fail CTS case is android.hardware.camera2.cts.RobustnessTest#testMandatoryOutputCombinations

Here is some more background for the issue. Our camera/ISP hardware supports 2 independent, parallel streams, but one of those two paths is limited to a maximum size of 720p. That path is used for generating the preview stream because we were only targeting 720p displays. During 1080p recording we are generating the recording output at 1080p with the one path that supports higher resolutions and the preview output with the path that is limited to 720p resolution. 

So if we don't take display size into account and 1080p preivew is requested, then we will report unsupported error.

And also in https://developer.android.com/reference/android/hardware/camera2/CameraDevice.html, we found this sentence "For the maximum size column, PREVIEW refers to the best size match to the device's screen resolution, or to 1080p (1920x1080), whichever is smaller." So we consider it is reasonable to consider display size not only in the place as this patch but also for PREVIEW_SIZE_BOUND in cts/tests/tests/hardware/src/android/hardware/camera2/cts/CameraTestUtils.java

Thanks,
Linda 

> Hi, Linda
 > 
 > We just tested this change on a small screen device and realize
 > there is a problem. Limiting the preview sizes here will also limit
 > the size available for preview callback, which is used by
 > application to do various kind of stuffs. Therefore, we think we
 > should probably not limited preview sizes by screen size. We could
 > possibly do that for the "preferred preview size for video" one
 > though.
 > 
 > Can you let us know which CTS tests will be failing on your device
 > without this change? I will check if there is alternative solution
 > for that. Thanks.

Patch-set: 5
Reviewer: Gerrit User 1075733 <1075733@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0