id3: Instead of aborting the parser on a zero-data frame, just skip that frame
Certain files seen in the wild, including the gb18030_5.mp3 file
used in CTS, include frames that consist of _just_ the header and
no data, and as such set the frame-size field to zero. That sample
file, for example, starts off with 2 empty PRIV frames, either of
which would abort the frame iterator and prevent the rest of the
(data-filled) tags to be read.

Fixes android.media.cts.MediaScannerTest#testEncodingDetection

Change-Id: I6c17af356791215d678d1e3d4a1f04883fae7537