SoftwareRenderer: Fix an incorrect sanity check
The mHeight/mWidth is set to the max value in adaptive video playback mode.
But the compared variable *size* is not the ABuffer's capacity but the length
of the valid data *mRangeLength* which is set by ABuffer::setRange() when the
FILL_BUFFER_DONE event occurs.

mCropHeight/mCropWidth is the real size of the display area which should never
be larger than the ABuffer's mRangeLength.

Change-Id: I1ada41d6ef5b10cc6ef9988c56086b30184c4880
Signed-off-by: Dan Liang <dan.liang@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>