Update patch set 1
Patch Set 1:

(1 comment)

> (2 comments)
 > 
 > I don't quite understand the need for this CL as vsyncsForLastFrame
 > is only logged, and is not used for any decisions.

The reason lies in limitations of the integer overflow sanitization as implemented. As it is right now, the generated code simply triggers a call to abort() upon performing an operation that results in an integer overflow. There is no concept of whether the value is used for any control flow decisions as it never gets the opportunity to do so. As a result, we get SIGABRTs triggered by overflow cases that are benign from an exploitation perspective.

Patch-set: 1