Include af.fast_mixer_max_warmup property
The FastMixer uses a hard-coded behavior that has been shown to
increase cold latency when DSPs have internal buffering. If set,
this property will reduce the maximum of warmup cycles used by
the FastMixer to reduce cold latency. If the property is not set,
the default value (10 cycles) will be used.

Change-Id: I3b030b050ba267a327179385e44659aa237100d0
Signed-off-by: Alejandro Ochoa <alejandro.ochoa@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>