Update patch set 1
Patch Set 1: Code-Review-2

Unfortunately, we are about to change how stagefright and camera are embedded in mediaserver, so this check will very soon not work correctly anymore.  

However, we'll track this issue and adapt your CL for the new arrangement - thank you for letting us know about the problem.

Patch-set: 1