Update patch set 1
Patch Set 1:

Patch Set 1:
I apologize, but I don't quite follow the events that lead to this deadlock. 
Is the flow:
1. Client app connects to camera
2. At the end of client connect call, we notify BatteryStats of camera being open
3. However, during connect call processing, the BatteryStats service has died for some reason (why is it dying?)
4. This leads to an immediate binderDied() call when we try to notify BatteryStats, on the same thread as the connect call.
Is that accurate?  If so, why do you see BatteryStats dying? As a service in the system service, if it dies, the system generally restarts the runtime.

Yes, this is the steps the deadlocked trigged. 
but we don't know why the BatteryStats dying, I check the AP log, there isn't such information. But after the Battery dying and before system restart it, there will has a timeslot the Binder transaction failed. so we need to keep the camera service avoid UIWDT.

Patch-set: 1