NuPlayerDriver: getCurrentPosition can report old value
If getCurrentPosition is called after a song has ended
in the background, it will report the last position
that was available when the application was visible in
the foreground (due to position is not updated while
Music application is in the background and mState being
STATE_PAUSED).

This will make the progressbar in Music application to not
show the end position when it is brought to the foreground.

Change-Id: Ibcee1f6094b53ea15849b9470fb7fd3c45d8a869
Signed-off-by: Markus Grape <markus.grape@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>