CameraService: extract initCameraStatus method
Extract the piece of code that initialize the camera status
to a separate function called initCameraStatus. The purpose of this
refactoring is to support external camera hotplug.  mCameraStates
is supposed to include not only built-in cameras but also the external
cameras. However, mCameraStatus will never be updated in current
implementation. Extracting a separate function will prepare us to
update the mCameraStatus in the hotplug notification.

Change-Id: I79989afba04218b6e559b3b0dc0cbec4ba5a0074
Signed-off-by: Bin Chen <bin.chen@linaro.org>