CameraService: hotplug support for external camera
Currently hotplug for external camera is not supported.
mCameraStatus is initialized during CameraService start
and thus contains only the camera id of built-in cameras.
When an external camera being plugged in, since its camera id
can't be found in the mCameraStatus, this hotplug notification
will be ignored by the CameraService and the Application.

The issue is solved by checking whether the event is a
valid new external camera plug in event. If it is, add the camera id
into mCameraStatus and set is initial status as NOT_PRESENT. Following
logic is the same as updating the status of built-in camera.

In addition, currently the number of cameras will not be updated since
camera start, so it didn't reflect the external cameras that can be
used. Solving this issue by adding mNumberOfExternalCameras and update
it on the hotplug notification.

Change-Id: I30ac738a6f83bfe4fce658398a3c1c1d7a53136d
Signed-off-by: Bin Chen <bin.chen@linaro.org>