CameraService: hotplug support for external camera
Currently hotplug for external camera is not supported.
mCameraStatus is initialized during CameraService start
and thus contains only the camera id of built-in cameras.
When an external camera being plugged in, since its camera id
can't be found in the mCameraStatus, this hotplug notification
will be ignored by the CameraService and the Application.

The issue is solved by checking whether the event is a
valid new external camera plug in event. If it is, add the camera id
into mCameraStatus and set its initial status as NOT_PRESENT. Following
logic is the same as updating the status of built-in cameras.

In addition, currently the number of cameras will not be updated after
camera service start, so it won't count in the external cameras and thus
user won't be able to use them. This issue is solved by adding
mNumberOfExternalCameras and it will be updated in hotplug notification.

Change-Id: I30ac738a6f83bfe4fce658398a3c1c1d7a53136d
Signed-off-by: Bin Chen <bin.chen@linaro.org>