Update patch set 2
Patch Set 2:

The in-band ringtone feature is tested with following setups as peer devices
1. PTS
2. CE4A tool v 15.5 and
3. Beats and Plantronics headsets

Configured music track is heard on the headset on incoming ring alert each time.

When tested with a headset which does not support in-band ringtone, default tone is heard on the headset, and selected tune is played out on the phone and no absurd behavior is observed.

Patch-set: 2
Reviewer: Gerrit User 1095543 <1095543@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0