Update patch set 1
Patch Set 1:

Yes this fixes a bug. 
The bug is reproduced only on tablet platform with pass-thru audio support in nuplayer(added by nvidia)
Steps:
1. Copy multiple ac3/dts files in sdcard.
2. Connect a HDMI receiver which supports ac3/dts passthru.
2. Play any ac3/dts file through android gallery app
3. Intermittently audio loss happens.

Rootcause: When one stream is playing, nuplayer tries to open output for other streams in background, mostly to keep playback chain ready for next stream(not completely sure why it opens output).

Issue happens when format for playing stream and stream getting opened in background are not same.

When APM gets getOutput() call it finds already active pass-thru output descriptor opened with different format and as a result it calls closeOutput() for this.
Once output is closed, audiotrack calls restoreTrack() but fails to restore as restoring is not supported for pass-thru.

Patch-set: 1
Reviewer: Gerrit User 1099424 <1099424@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0