Update patch set 1
Patch Set 1:

adding a lock is not going to prevent the WV engine from accessing the file descriptor and modifying its position, is it? Are you saying that *while* the call to FileSource::DrmInitialization is in progress, something else is accessing the file descriptor at the same time?

Patch-set: 1
Reviewer: Gerrit User 1001525 <1001525@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0