stagefright: mpegps: Correct timestamp logic
For some of .mpg files we will get timestamps along
with sequence header. Previously we are sending this
timestamp in first buffer after getting sequence header
and for the remaining buffers we are sending zero timestamp.
Corrected the logic to send the set of frames (from sequence
header to sequence header) with the same timestamp.

Change-Id: Ib75fa94e28fa89b612bdd6481d8444c98fd13e28