Update patch set 1
Patch Set 1:

The AutoConditionLock is an automatic lock - when created, it acquires mServiceLock. When it is destroyed (on line 1586), it releases mServiceLock.   

If mServiceLock is already released, the automatic release on 1586 will cause an error. So any manual unlocking within the scope of existence of the AutoConditionLock must be paired with relocking.  

Generally we try hard never to have to unlock manually, since it's so error-prone, but this is one of the places where it was necessary.

Patch-set: 1
Reviewer: Gerrit User 1010050 <1010050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0