Codecs: Fix EOS handling in soft codecs
When EOS is sent along with non-zero length input buffer and if decoder
was still not producing output, due to initial latency, then decoder
would not produce any output. For eg. When only one input buffer with EOS
set is sent to the decoder, decoder was not producing any output.

Applies for SoftHEVC, SoftAVC and SoftMpeg2 decoders.

Bug: 28965335