Update patch set 1
Patch Set 1:

Ouch! Reducing a java app into cpp snippet is not trivial. Good luck! 

In the meantime, if you guys could provide the neon sequence, that'd help. 

Finally, if this is related to floating point IEEE problems, I have fixed clang behaviour to not vectorise FP on ARM unless fast math is chosen, which may help, but only if you use a recent clang. 

Hope that helps.

Patch-set: 1