Update patch set 1
Patch Set 1:

> I confirmed that your patch (https://android-review.googlesource.com/249846)
 > fixes the issue.

Thank you very much for testing, and for letting us know about this issue!  

We won't be able to merge that fix into the N release, but it will be in the first available followup release from that. In the meanwhile, please feel free to include it in your own builds.

Patch-set: 1
Reviewer: Gerrit User 1010050 <1010050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0