Update patch set 1
Patch Set 1:

Please do not merge functional changes to the camera service without review from someone from the camera framework team.

Did you actually test this? ValidateClientPermissionsLocked does a number of checks that call to services outside of the camera service, which to my understanding are not present on a Brillo platform.  

This will at the minimum produce a bunch of spurious error logging that is misleading.

Please revert.

Patch-set: 1