Update patch set 2
Patch Set 2:

> etalvala, yinchiayeh: In case either of you missed this, here's the
 > doc for the new process for adding NDK headers/libs:
 > https://android.googlesource.com/platform/ndk/+/master/docs/PlatformApis.md
 > 
 > The version script was generated automatically from the
 > libcamera2ndk symbol files in development/ndk.
 > 
 > I could use some help with the headers though. We need to make sure
 > the headers can be included regardless of the target API level, so
 > we need to add `#if __ANDROID_API__ >= 24` where necessary. Are all
 > the headers used by libcamera2ndk.so new with android-24, or did
 > they also provide other APIs for lower platform levels?

All camera NDK APIs are new in 24 (N). We do not have any NDK API prior to 24.

On the other hand, this is not the case for media NDK APIs (some of them are shipped in L/L-MR1, but I also added AImageReader APIs in 24)

Patch-set: 2
Label: Verified=0