Update patch set 1
Patch Set 1:

Hi, Dan

Could you please help to check if this is the right solution for such problem?
I check the logs and found that you enabled most of the sanitizer for integer, and also have some several ways for the abort problem, like disable the sanitizer again in android.mk, or disable sanitizer for specific functions  like this, or changing some source code to avoid this abort problem. I am not sure which one is best, and from the commit log, seems this method is used most, so I submitted this change for review, please help to give comments on this and do the check .

Thanks,
Yongqin Liu

Patch-set: 1
Reviewer: Gerrit User 1041813 <1041813@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0