Update patch set 7
Patch Set 7:

(10 comments)

Hi Jerry, thank you for handling lots of comments so far.
I think we are close to the end after solving race condition around MtpFfsHandle.
Could you ask +2 for someone in the framework team? I'm not a member of Android framework team and would like to defer +2 for someone in the framework team.

Patch-set: 7