Decode all AMR-WB bits in MIME/storage file format
The MIME/storage file format (IETF RFC 4867 section 5.3) requires to
decode all speech bits of a frame block. Here, the code divides a frame
by 8 and 2 to convert not each speech bit individually but four bits at
a time. This devision creates a remainder of 0 to 7 speech bits. In case
of AMR-WB, the remaining speech bits depend on the current mode (like
12.65 or 23.85 kbit/s) and are 0 to 5 bits.

The previous code expected just a single remainder bit. Even that was
not decoded, because each octet was copied into a temporary variable.
Therefore, up to five bits got lost per frame. Furthermore because of
this, Silence Insertion Descriptor (SID) frames were detected not as
RX_SID_UPDATE but always as RX_SID_FIRST. This created noise when Voice
Activation Detection (VAD) was enabled in the encoder.

identified downstream by Morten Tryfoss via the OpenCORE-AMR project

Test: manual - placing syslog messages into code
Change-Id: I2c1db946e7aed1bf5963a73e83ccff7e190f4c4c
Signed-off-by: Alexander Traud <pabstraud@compuserve.com>