Update patch set 4
Patch Set 4:

In terms of levels of access, we distinguish not whether it's a framework component, but where it's a native daemon, system_server, priv-app, normal app (aka untrusted_app in SELinux terms). For example, MtpServer code is in frameworks/, but the code runs in a priv-app MediaProvider. The aim here is to start restricting access from any Android package (priv-app or normal app) to device identifiers not resettable by user and making privileged access go through special funnels exposed via public or @SystemApi of framework.

Patch-set: 4