Update patch set 10
Patch Set 10: Code-Review+1

This is not my code base. I would like to defer to the owners or recent committers of this file to decide if any testing or changes in return code and its handling is necessary. The logical change to address the security bug looks good to me.

Patch-set: 10
Reviewer: Gerrit User 1118893 <1118893@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1