Update patch set 3
Patch Set 3:

Yes, at this time the integer variants are not called (without enabling the integer mixer).  However, in the future, the resampler may be called outside of AF (potentially on the client side). I just wanted to follow up on your earlier comment that you were going to do it, so we can make sure to get it in if you had it ready :-).  No worries for now.

Patch-set: 3