Update patch set 1
Patch Set 1:

(1 comment)

The suggestion does not work as is because of another error. 
I can make it dynamic sized, but then it needs to be zero initialized with memset. But maybe this is probably too much of change for a simple warning fix.

Patch-set: 1
Reviewer: Gerrit User 1130573 <1130573@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0