Update patch set 1
Patch Set 1:

(2 comments)

This commit was a result of analysis of "dumpsys" where the frame-count was seen very high. I somehow wanted it to be small enough to avoid immediate or long term stall (in either client or server of track).

So I thought of using frame-count from HAL and then multiply the highest frame-count between playback and record with 2 to ensure that, either client or server of the patch track has space to dump the frames (at-least 1 more than required).

Few options -
* We can add check to see if playback supports fast-mixer or not and based on that, decide to use normal frame-count or HAL's frame-count.
OR
* Since normal frame-count is multiple of HAL's frame-count (multiplier being between 1.0 to 2.0), we simply use normal frame-count always for playback.

I think the point I am trying to make is - to come up with something which will ensure neither track's client or server is stalled due to unavailability of space in track buffer.
Please suggest any other option that you might think of.

Thanks.

Patch-set: 1
Label: Verified=0