Update patch set 1
Patch Set 1:

"The
number of packets an endpoint on a device can send or receive at a time without an intermediate
ACK TP is reported by the device in the endpoint companion descriptor (refer to Section 9.6.7) for
that endpoint. An endpoint that reports more than one packet in its maximum burst size is
considered to be able to support “Burst” Transactions." From usb 3.1 spec. 

So bMaxBurst only affects performance, not compatibility or functionality. As for why the maximum value isn't the fastest, I'm not quite sure. I tried values between 2 and 15 and this seems like the best for marlin, but it is possible that newer usb 3 devices would be more optimized with a different value. One possibility is that for very large bursts the packets must be buffered since the device cannot retrieve them fast enough and thus is a little slower.

Patch-set: 1
Reviewer: Gerrit User 1113594 <1113594@85c56323-6fa9-3386-8a01-6480fb634889>