Update patch set 2
Patch Set 2:

If frameworks is not disconnecting the surface upon release, we should patch frameworks to do so. I think this would be more robust than trying to disconnect on every error scenario implicitly. We're better off to wait for the app to call release and clean up everything (not just the surface).

Patch-set: 2
Reviewer: Gerrit User 1039918 <1039918@85c56323-6fa9-3386-8a01-6480fb634889>