support ashmem for FileSource
If fd was /dev/ashmem, Use ashmem_get_size_region to get size
instead of fstat. The misc device of /dev/ashmem does not support
fstat. Therefore, if we try to play a file in ashmem with SoundPool,
FileSource will adjust the length/offset to 0/0, which will result
in extract fail. Anyway, it fail to play in Android N, which success
with Android M.

Test: Play audio file which stored in ashmem with SoundPool API.
Change-Id: I383de43c5d260e08d1cb2541fe6dfdb13632de0d
Signed-off-by: zengjing <zengjing@xiaomi.com>