Do not apply volume ramp in initial volume seting
There are two separate threads during playback of a track. First one is
applying volume and the second one is changing a state of playback.
After a track is created it is possible that a second thread will change
track's state to pausing before a volume was initialized. This caused a
first thread to apply volume ramp.
This CL is change not to apply volume ramp in initial volume
setting.

Bug: XXX
Test: Check shutter sound in video recording

Change-Id: I7bd876eeb7a84d7ac14b7d03db0571e36c077d63