Improve the issue of taking long time to prepare long duration MIDI
It takes long time to prepare long duration MIDI. When MIDI file is
parsed, MidiExtractor uses DataSource::readAt() many times. This
overhead is much bigger because interprocess communication happens when
DataSource::readAt() is called. To improve this, this fix adds
to change cache size in TinyDataSource. The cache size of default
implementation is 2K. It is not enough in MidiExtractor. Si file size
is handled as cache size in MidiExtractor from experimenting.

Bug: 67480585
Test: Play a long time dureation MIDI file.
Change-Id: I5455f0974bc14c4e9f650f29713822fbdbd86c34