Reduce number of skipped frames before choosing thumbnail
Till now 20'th key frame was picked as thumbnail for matroska videos.
Probably it was done that way in order to avoid "black thumbnails" which
often appear on beginning of the video. When there are less frames then
last one will be used instead of 20'th. However this has some drawback
because if we have less than 20 key frames whole content will be parsed
to find them and it requires getting whole content.

Now if media extractor is in other process, the content
will be transferred through binder causing noticable delays.

Test: check photos with "webm + less sync point" contents

Change-Id: I24b4a58b892aa0d27bdd18ce28dbcab9aab9b035